refactor : move dom code to adapter
