+++ pages/functional/factorial.html
@@ -63,6 +63,6 @@
       <p><a href="/">Back to Homepage</a></p>
     </main>
 
-    <script type="module" src="./js/factorial.js"></script>
+    <script type="module" src="./adapter/factorial.js"></script>
   </body>
 </html>
