+++ pages/functional/fizz-buzz.html
@@ -35,6 +35,6 @@
       <div class="result-container-fizzbuzz" id="result"></div>
     </main>
 
-    <script type="module" src="./js/fizz-buzz.js"></script>
+    <script type="module" src="./adapter/fizz-buzz.js"></script>
   </body>
 </html>
