+++ pages/functional/js/factorial.js
@@ -50,18 +50,8 @@ function countFactorial(n, method) {
   }
 }
 
-document.getElementById("form").addEventListener("submit", function (event) {
-  event.preventDefault();
-
-  try {
-    const n = event.target["n"].value;
-    const method = event.target["method"].value;
-
-    const result = countFactorial(n, method);
-
-    document.getElementById("result").textContent = result;
-  } catch (error) {
-    alert(error.message);
-    console.error(error);
-  }
-});
+export {
+  countFactorialUsingLoop,
+  countFactorialUsingRecursive,
+  countFactorial,
+};
