+++ pages/functional/js/fibonacci.js
@@ -67,20 +67,8 @@ function generateFibonacci(sequence, method) {
   }
 }
 
-document.getElementById("form").addEventListener("submit", function (event) {
-  event.preventDefault();
-
-  try {
-    const sequence = event.target["sequence"].value;
-    const method = event.target["method"].value;
-
-    const numbers = generateFibonacci(sequence, method);
-
-    document.getElementById("result").innerHTML = numbers
-      .map(number => `<div class="result-item-fibonacci">${number}</div>`)
-      .join("");
-  } catch (error) {
-    alert(error.message);
-    console.error(error);
-  }
-});
+export {
+  generateFibonacciUsingLoop,
+  generateFibonacciUsingRecursive,
+  generateFibonacci,
+};
