+++ pages/functional/js/palindrome.js
@@ -89,18 +89,9 @@ function generatePalindromeStatus(word, method) {
     : "No, this word is not a palindrome.";
 }
 
-document.getElementById("form").addEventListener("submit", function (event) {
-  event.preventDefault();
-
-  try {
-    const word = event.target["word"].value;
-    const method = event.target["method"].value;
-
-    const result = generatePalindromeStatus(word, method);
-
-    document.getElementById("result").textContent = result;
-  } catch (error) {
-    alert(error.message);
-    console.error(error);
-  }
-});
+export {
+  isPalindromeUsingLoop,
+  isPalindromeUsingRecursive,
+  isPalindromeUsingReverse,
+  generatePalindromeStatus,
+};
