+++ pages/oop/fibonacci.html
@@ -62,6 +62,6 @@
       <p><a href="/">Back to Homepage</a></p>
     </main>
 
-    <script type="module" src="./js/Fibonacci.js"></script>
+    <script type="module" src="./adapter/fibonacci.js"></script>
   </body>
 </html>
