+++ pages/oop/js/Factorial.js
@@ -56,19 +56,3 @@ export class Factorial {
     }
   }
 }
-
-document.getElementById("form").addEventListener("submit", function (event) {
-  event.preventDefault();
-
-  try {
-    const n = event.target["n"].value;
-    const method = event.target["method"].value;
-
-    const result = new Factorial(n).count(method);
-
-    document.getElementById("result").textContent = result;
-  } catch (error) {
-    alert(error.message);
-    console.error(error);
-  }
-});
