+++ pages/oop/js/FizzBuzz.js
@@ -35,29 +35,3 @@ export class FizzBuzz {
     return result;
   }
 }
-
-document.getElementById("form").addEventListener("submit", function (event) {
-  event.preventDefault();
-
-  try {
-    const sequence = event.target["sequence"].value;
-
-    const fizzBuzzs = new FizzBuzz(sequence).generate();
-
-    document.getElementById("result").innerHTML = fizzBuzzs
-      .map(
-        (fizzBuzz, index) =>
-          `<div class="result-item-fizzbuzz bg-${
-            {
-              "fizz buzz": "red",
-              fizz: "blue",
-              buzz: "green",
-            }[fizzBuzz]
-          }">${index + 1} ${fizzBuzz}</div>`,
-      )
-      .join("");
-  } catch (error) {
-    alert(error.message);
-    console.error(error);
-  }
-});
