+++ pages/oop/js/Palindrome.js
@@ -95,19 +95,3 @@ export class Palindrome {
       : "No, this word is not a palindrome.";
   }
 }
-
-document.getElementById("form").addEventListener("submit", function (event) {
-  event.preventDefault();
-
-  try {
-    const word = event.target["word"].value;
-    const method = event.target["method"].value;
-
-    const result = new Palindrome(word).generateStatus(method);
-
-    document.getElementById("result").textContent = result;
-  } catch (error) {
-    alert(error.message);
-    console.error(error);
-  }
-});
