+++ pages/oop/palindrome.html
@@ -65,6 +65,6 @@
       <p><a href="/">Back to Homepage</a></p>
     </main>
 
-    <script type="module" src="./js/Palindrome.js"></script>
+    <script type="module" src="./adapter/palindrome.js"></script>
   </body>
 </html>
